+++ jwst/resample/gwcs_drizzle.py
@@ -144,29 +144,29 @@ class GWCSDrizzle:
             Must have the same dimensions as insci. If none is supplied,
             the weighting is set to one.
 
-        xmin : float, optional
+        xmin : int, optional
             This and the following three parameters set a bounding rectangle
-            on the output image. Only pixels on the output image inside this
-            rectangle will have their flux updated. Xmin sets the minimum value
-            of the x dimension. The x dimension is the dimension that varies
-            quickest on the image. If the value is zero or less, no minimum will
-            be set in the x dimension. All four parameters are zero based,
-            counting starts at zero.
+            on the input image. Only pixels on the input image inside this
+            rectangle will have their flux added to the output image. Xmin
+            sets the minimum value of the x dimension. The x dimension is the
+            dimension that varies quickest on the image. All four parameters
+            are zero based, counting starts at zero.
 
-        xmax : float, optional
+        xmax : int, optional
             Sets the maximum value of the x dimension on the bounding box
-            of the output image. If the value is zero or less, no maximum will
-            be set in the x dimension.
+            of the input image. If ``xmax = 0``, no maximum will
+            be set in the x dimension (all pixels in a row of the input image
+            will be resampled).
 
-        ymin : float, optional
+        ymin : int, optional
             Sets the minimum value in the y dimension on the bounding box. The
             y dimension varies less rapidly than the x and represents the line
-            index on the output image. If the value is zero or less, no minimum
-            will be set in the y dimension.
+            index on the input image.
 
-        ymax : float, optional
-            Sets the maximum value in the y dimension. If the value is zero or
-            less, no maximum will be set in the y dimension.
+        ymax : int, optional
+            Sets the maximum value in the y dimension. If ``ymax = 0``,
+            no maximum will be set in the y dimension (all pixels in a column
+            of the input image will be resampled).
 
         expin : float, optional
             The exposure time of the input image, a positive number. The
@@ -279,31 +279,29 @@ def dodrizzle(insci, input_wcs, inwht, output_wcs, outsci, outwht, outcon,
         this function is called and incremented by one on each subsequent
         call.
 
-    xmin : float, optional
+    xmin : int, optional
         This and the following three parameters set a bounding rectangle
         on the input image. Only pixels on the input image inside this
         rectangle will have their flux added to the output image. Xmin
         sets the minimum value of the x dimension. The x dimension is the
-        dimension that varies quickest on the image. If the value is zero,
-        no minimum will be set in the x dimension. All four parameters are
-        zero based, counting starts at zero.
+        dimension that varies quickest on the image. All four parameters
+        are zero based, counting starts at zero.
 
-    xmax : float, optional
+    xmax : int, optional
         Sets the maximum value of the x dimension on the bounding box
-        of the input image. If the value is zero, no maximum will
-        be set in the x dimension, the full x dimension of the output
-        image is the bounding box.
+        of the input image. If ``xmax = 0``, no maximum will
+        be set in the x dimension (all pixels in a row of the input image
+        will be resampled).
 
-    ymin : float, optional
+    ymin : int, optional
         Sets the minimum value in the y dimension on the bounding box. The
         y dimension varies less rapidly than the x and represents the line
-        index on the input image. If the value is zero, no minimum  will be
-        set in the y dimension.
+        index on the input image.
 
-    ymax : float, optional
-        Sets the maximum value in the y dimension. If the value is zero, no
-        maximum will be set in the y dimension,  the full x dimension
-        of the output image is the bounding box.
+    ymax : int, optional
+        Sets the maximum value in the y dimension. If ``ymax = 0``,
+        no maximum will be set in the y dimension (all pixels in a column
+        of the input image will be resampled).
 
     pixfrac : float, optional
         The fraction of a pixel that the pixel flux is confined to. The